Added the Failure exception and the 'is_forwarded' property.
The processing of failed operations has changed, a Failure
exception is now thrown when the operation's response has a
client or server error. The exception might be rescued by the
operation dispatcher in the following cases:

- The request is an XHR: the response of the operation is
returned.

- The operation was "forwarded" (see below): the exception is
discarted and the dispatch of the request continue.

An operation is considered "forwarded" when the actual destination
and operation type is defined using the request parameters
DESTINATION and NAME. The URI of the request is irrelevant to
forwarded operation, and whether they succeed or fail the
dispatch process simply continues. This allows forms to be
posted to their display URL (not the URL of the operation) and
displayed again if an error occurs.